feat: add system testing quality gate (#112)
* feat: add system testing quality gate (#0008)

- Add per-handler xUnit test files across all services (195 tests total)
- Rewrite saga tests using MassTransit InMemoryTestHarness
- Extract per-domain test fixtures and shared testing infrastructure
- Add TESTING.md with TDD workflow, one-file-per-handler rule, and worked example
- Move PaymentService tests from src/ to tests/ directory


* feat: add test suite, coverage tooling, and service CQRS migrations

- Add comprehensive unit tests across all services (ApiGateway, Catalog, Identity, Media, Notification, Order, Payment, UserService)
- Add coverage.ps1, coverage.runsettings, and show-coverage skill for quality gate
- Migrate UserService from MVC controllers to CQRS/Minimal API pattern
- Restructure MediaService persistence layer (move Infrastructure/Data to Persistence)
- Relocate IdentityService DataSeeder and IdentityEventPublisher to canonical paths
- Modify BaseRepository with updated query/pagination methods
- Update done-story skill and agent docs with testing conventions


* Refactor code structure for improved readability and maintainability

* fix: address code review findings from PR #112

- Remove committed coverage-results XML artifact; add /coverage-results/ to .gitignore
- Apply explicit AdminOrUser policy to UserAddressEndpoints group for consistency with UserEndpoints
- Return 404 from GetDefaultUserAddress when no default address exists
- Fix quality-gate coverage XML selection to use most-recently-written file on retry runs


* feat: enhance start-story documentation with TDD ordering and test failure protocol

* fix: remove ExcludeFromCodeCoverage from domain constructors and restore logic guards

- Remove [ExcludeFromCodeCoverage] and unused using directives from private
  parameterless constructors across CatalogService, PaymentService, and
  UserService domain classes; also remove from Payment.ThrowIfExpired() and
  PhoneNumber.FormatForDisplay() private methods
- Restore three early-return guards in CreateOrderConsumer.ApplyPlatformCouponProration
  to prevent division-by-zero when no weighted orders exist
- Restore null check before ApplyCoupon in GetCartSummaryQueryHandler to
  prevent NullReferenceException when a platform coupon code is not found


---------

+++ src/HiveSpace.OrderService/HiveSpace.OrderService.Api/Consumers/Sync/ProductRefSyncConsumer.cs
@@ -39,10 +39,10 @@ public class ProductRefSyncConsumer(OrderDbContext db, ILogger<ProductRefSyncCon
     public async Task Consume(ConsumeContext<ProductSkuUpdatedIntegrationEvent> context)
     {
         var msg = context.Message;
-        var existing = await db.SkuRefs.FindAsync([msg.SkuId], context.CancellationToken);
+        var existing = await db.SkuRefs.FindAsync([(long)msg.SkuId], context.CancellationToken);
         if (existing is null)
         {
-            db.SkuRefs.Add(new SkuRef(msg.SkuId, msg.ProductId, msg.SkuNo, msg.Price, msg.Currency,
+            db.SkuRefs.Add(new SkuRef((long)msg.SkuId, (long)msg.ProductId, msg.SkuNo, msg.Price, msg.Currency,
                 imageUrl: msg.ThumbnailUrl, attributes: null, skuName: msg.SkuName));
             logger.LogInformation("SkuRef created. SkuId={SkuId}", msg.SkuId);
         }
